feat(sdds-alibs/sdds.serv): separate button builders and rearrange imports
