+++ playground/sandbox-compose/src/main/kotlin/com/sdds/playground/sandbox/core/Properties.kt
@@ -25,8 +25,8 @@ import com.sdds.compose.uikit.IconButton
 import com.sdds.compose.uikit.Switch
 import com.sdds.compose.uikit.Text
 import com.sdds.playground.sandbox.R
-import com.sdds.serv.styles.buttons.Clear
-import com.sdds.serv.styles.buttons.Xs
+import com.sdds.serv.styles.button.icon.Clear
+import com.sdds.serv.styles.button.icon.Xs
 import com.sdds.serv.tokens.compose.SddsServTheme
 import com.sdds.icons.R.drawable as Icons
