--- tokens/plasma.sd.service.compose/src/main/kotlin/com/sdds/plasma/sd/service/styles/TextAreaStyles.kt
@@ -1,228 +1,0 @@
-package com.sdds.plasma.sd.service.styles
-
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.unit.dp
-import com.sdds.compose.uikit.ChipGroup
-import com.sdds.compose.uikit.EmbeddedChip
-import com.sdds.compose.uikit.ScrollBar
-import com.sdds.compose.uikit.TextArea
-import com.sdds.compose.uikit.TextField
-import com.sdds.compose.uikit.TextFieldStyle
-import com.sdds.compose.uikit.TextFieldStyleBuilder
-import com.sdds.compose.uikit.adjustBy
-import com.sdds.plasma.sd.service.tokens.compose.PlasmaSdServiceTheme
-
-val TextArea.Xs: TextFieldStyleBuilder
-    @Composable
-    get() = TextFieldStyle.builder(this)
-        .fieldAppearance(TextField.FieldAppearance.Solid)
-        .shape(PlasmaSdServiceTheme.shapes.roundS)
-        .helperTextPlacement(TextField.HelperTextPlacement.Inner)
-        .chipGroupStyle(
-            ChipGroup.Dense.chipStyle(
-                EmbeddedChip.Xs.Secondary.style(),
-            ).style(),
-        )
-        .scrollBar(scrollBar)
-        .dimensions(
-            TextField.Dimensions(
-                boxPaddingStart = 8.dp,
-                boxPaddingEnd = 8.dp,
-                boxPaddingTopInnerLabel = 8.dp,
-                boxPaddingBottomInnerLabel = 8.dp,
-                boxPaddingTopOuterLabel = 8.dp,
-                boxPaddingBottomOuterLabel = 8.dp,
-                innerLabelPadding = 0.dp,
-                outerLabelPadding = 6.dp,
-                optionalPadding = 4.dp,
-                helperTextPaddingInner = 8.dp,
-                helperTextPaddingOuter = 4.dp,
-                startContentEndPadding = 4.dp,
-                endContentStartPadding = 4.dp,
-                chipsPadding = 6.dp,
-                boxMinHeight = 32.dp,
-                alignmentLineHeight = 32.dp,
-                iconSize = 16.dp,
-                indicatorDimensions = TextField.Dimensions.IndicatorDimensions(
-                    startLabelHorizontalPadding = 4.dp,
-                    startLabelVerticalPadding = 0.dp,
-                    endLabelHorizontalPadding = 4.dp,
-                    endLabelVerticalPadding = 2.dp,
-                    fieldIndicatorSize = 6.dp,
-                    labelIndicatorSize = 6.dp,
-                ),
-            ),
-        )
-        .innerLabelStyle(PlasmaSdServiceTheme.typography.bodyXxsNormal)
-        .outerLabelStyle(PlasmaSdServiceTheme.typography.bodyXsNormal)
-        .innerOptionalStyle(PlasmaSdServiceTheme.typography.bodyXxsNormal)
-        .outerOptionalStyle(PlasmaSdServiceTheme.typography.bodyXsNormal)
-        .valueStyle(PlasmaSdServiceTheme.typography.bodyXsNormal)
-        .captionStyle(PlasmaSdServiceTheme.typography.bodyXsNormal)
-        .counterStyle(PlasmaSdServiceTheme.typography.bodyXsNormal)
-        .placeholderStyle(PlasmaSdServiceTheme.typography.bodyXsNormal)
-        .dropInnerLabel(true)
-        .singleLine(false)
-
-val TextArea.S: TextFieldStyleBuilder
-    @Composable
-    get() = TextFieldStyle.builder(this)
-        .fieldAppearance(TextField.FieldAppearance.Solid)
-        .shape(PlasmaSdServiceTheme.shapes.roundM.adjustBy(all = (-2).dp))
-        .helperTextPlacement(TextField.HelperTextPlacement.Inner)
-        .chipGroupStyle(
-            ChipGroup.Dense.chipStyle(
-                EmbeddedChip.S.Secondary.style(),
-            ).style(),
-        )
-        .scrollBar(scrollBar)
-        .dimensions(
-            TextField.Dimensions(
-                boxPaddingStart = 12.dp,
-                boxPaddingEnd = 12.dp,
-                boxPaddingTopInnerLabel = 4.dp,
-                boxPaddingBottomInnerLabel = 12.dp,
-                boxPaddingTopOuterLabel = 8.dp,
-                boxPaddingBottomOuterLabel = 12.dp,
-                innerLabelPadding = 0.dp,
-                outerLabelPadding = 8.dp,
-                optionalPadding = 4.dp,
-                helperTextPaddingInner = 12.dp,
-                helperTextPaddingOuter = 4.dp,
-                startContentEndPadding = 4.dp,
-                endContentStartPadding = 6.dp,
-                chipsPadding = 6.dp,
-                boxMinHeight = 40.dp,
-                alignmentLineHeight = 40.dp,
-                iconSize = 24.dp,
-                indicatorDimensions = TextField.Dimensions.IndicatorDimensions(
-                    startLabelHorizontalPadding = 4.dp,
-                    startLabelVerticalPadding = 0.dp,
-                    endLabelHorizontalPadding = 4.dp,
-                    endLabelVerticalPadding = 4.dp,
-                    fieldIndicatorSize = 6.dp,
-                    labelIndicatorSize = 6.dp,
-                ),
-            ),
-        )
-        .innerLabelStyle(PlasmaSdServiceTheme.typography.bodyXsNormal)
-        .outerLabelStyle(PlasmaSdServiceTheme.typography.bodySNormal)
-        .innerOptionalStyle(PlasmaSdServiceTheme.typography.bodyXsNormal)
-        .outerOptionalStyle(PlasmaSdServiceTheme.typography.bodySNormal)
-        .valueStyle(PlasmaSdServiceTheme.typography.bodySNormal)
-        .captionStyle(PlasmaSdServiceTheme.typography.bodyXsNormal)
-        .counterStyle(PlasmaSdServiceTheme.typography.bodyXsNormal)
-        .placeholderStyle(PlasmaSdServiceTheme.typography.bodySNormal)
-        .singleLine(false)
-
-val TextArea.M: TextFieldStyleBuilder
-    @Composable
-    get() = TextFieldStyle.builder(this)
-        .fieldAppearance(TextField.FieldAppearance.Solid)
-        .shape(PlasmaSdServiceTheme.shapes.roundM)
-        .helperTextPlacement(TextField.HelperTextPlacement.Inner)
-        .chipGroupStyle(
-            ChipGroup.Dense.chipStyle(
-                EmbeddedChip.M.Secondary.style(),
-            ).style(),
-        )
-        .scrollBar(scrollBar)
-        .dimensions(
-            TextField.Dimensions(
-                boxPaddingStart = 14.dp,
-                boxPaddingEnd = 14.dp,
-                boxPaddingTopInnerLabel = 6.dp,
-                boxPaddingBottomInnerLabel = 12.dp,
-                boxPaddingTopOuterLabel = 12.dp,
-                boxPaddingBottomOuterLabel = 12.dp,
-                innerLabelPadding = 2.dp,
-                outerLabelPadding = 10.dp,
-                optionalPadding = 4.dp,
-                helperTextPaddingInner = 12.dp,
-                helperTextPaddingOuter = 4.dp,
-                startContentEndPadding = 6.dp,
-                endContentStartPadding = 8.dp,
-                chipsPadding = 6.dp,
-                boxMinHeight = 48.dp,
-                alignmentLineHeight = 48.dp,
-                iconSize = 24.dp,
-                indicatorDimensions = TextField.Dimensions.IndicatorDimensions(
-                    startLabelHorizontalPadding = 6.dp,
-                    startLabelVerticalPadding = 0.dp,
-                    endLabelHorizontalPadding = 4.dp,
-                    endLabelVerticalPadding = 4.dp,
-                    fieldIndicatorSize = 8.dp,
-                    labelIndicatorSize = 6.dp,
-                ),
-            ),
-        )
-        .innerLabelStyle(PlasmaSdServiceTheme.typography.bodyXsNormal)
-        .outerLabelStyle(PlasmaSdServiceTheme.typography.bodyMNormal)
-        .innerOptionalStyle(PlasmaSdServiceTheme.typography.bodyXsNormal)
-        .outerOptionalStyle(PlasmaSdServiceTheme.typography.bodyMNormal)
-        .valueStyle(PlasmaSdServiceTheme.typography.bodyMNormal)
-        .captionStyle(PlasmaSdServiceTheme.typography.bodyXsNormal)
-        .counterStyle(PlasmaSdServiceTheme.typography.bodyXsNormal)
-        .placeholderStyle(PlasmaSdServiceTheme.typography.bodyMNormal)
-        .singleLine(false)
-
-val TextArea.L: TextFieldStyleBuilder
-    @Composable
-    get() = TextFieldStyle.builder(this)
-        .fieldAppearance(TextField.FieldAppearance.Solid)
-        .shape(PlasmaSdServiceTheme.shapes.roundM.adjustBy(all = 2.dp))
-        .helperTextPlacement(TextField.HelperTextPlacement.Inner)
-        .chipGroupStyle(
-            ChipGroup.Dense.chipStyle(
-                EmbeddedChip.L.Secondary.style(),
-            ).style(),
-        )
-        .scrollBar(scrollBar)
-        .dimensions(
-            TextField.Dimensions(
-                boxPaddingStart = 16.dp,
-                boxPaddingEnd = 16.dp,
-                boxPaddingTopInnerLabel = 9.dp,
-                boxPaddingBottomInnerLabel = 12.dp,
-                boxPaddingTopOuterLabel = 16.dp,
-                boxPaddingBottomOuterLabel = 12.dp,
-                innerLabelPadding = 2.dp,
-                outerLabelPadding = 12.dp,
-                optionalPadding = 4.dp,
-                helperTextPaddingInner = 12.dp,
-                helperTextPaddingOuter = 4.dp,
-                startContentEndPadding = 8.dp,
-                endContentStartPadding = 10.dp,
-                chipsPadding = 6.dp,
-                boxMinHeight = 56.dp,
-                alignmentLineHeight = 56.dp,
-                iconSize = 24.dp,
-                indicatorDimensions = TextField.Dimensions.IndicatorDimensions(
-                    startLabelHorizontalPadding = 6.dp,
-                    startLabelVerticalPadding = 0.dp,
-                    endLabelHorizontalPadding = 4.dp,
-                    endLabelVerticalPadding = 4.dp,
-                    fieldIndicatorSize = 8.dp,
-                    labelIndicatorSize = 6.dp,
-                ),
-            ),
-        )
-        .innerLabelStyle(PlasmaSdServiceTheme.typography.bodyXsNormal)
-        .outerLabelStyle(PlasmaSdServiceTheme.typography.bodyLNormal)
-        .innerOptionalStyle(PlasmaSdServiceTheme.typography.bodyXsNormal)
-        .outerOptionalStyle(PlasmaSdServiceTheme.typography.bodyLNormal)
-        .valueStyle(PlasmaSdServiceTheme.typography.bodyLNormal)
-        .captionStyle(PlasmaSdServiceTheme.typography.bodyXsNormal)
-        .counterStyle(PlasmaSdServiceTheme.typography.bodyXsNormal)
-        .placeholderStyle(PlasmaSdServiceTheme.typography.bodyLNormal)
-        .singleLine(false)
-
-private val scrollBar: ScrollBar
-    @Composable
-    get() = ScrollBar(
-        indicatorThickness = 1.dp,
-        indicatorColor = PlasmaSdServiceTheme.colors.surfaceDefaultTransparentTertiary,
-        backgroundColor = PlasmaSdServiceTheme.colors.surfaceDefaultTransparentPrimary,
-        padding = PaddingValues(top = 18.dp, end = 2.dp, bottom = 36.dp),
-    )
