+++ tokens/plasma.sd.service.compose/src/main/kotlin/com/sdds/plasma/sd/service/styles/avatar/AvatarStyles.kt
@@ -1,4 +1,4 @@
-package com.sdds.plasma.sd.service.styles
+package com.sdds.plasma.sd.service.styles.avatar
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.geometry.Offset
