+++ tokens/plasma.sd.service.compose/src/main/kotlin/com/sdds/plasma/sd/service/styles/button/icon/IconButtonStyles.kt
@@ -1,4 +1,4 @@
-package com.sdds.plasma.sd.service.styles
+package com.sdds.plasma.sd.service.styles.button.icon
 
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.runtime.Composable
