+++ tokens/plasma.sd.service.compose/src/main/kotlin/com/sdds/plasma/sd/service/styles/chip/group/ChipGroupStyles.kt
@@ -1,4 +1,4 @@
-package com.sdds.plasma.sd.service.styles
+package com.sdds.plasma.sd.service.styles.chip.group
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.unit.dp
