+++ tokens/plasma.sd.service.compose/src/main/kotlin/com/sdds/plasma/sd/service/styles/textfield/TextAreaStyles.kt
@@ -1,37 +1,48 @@
-package com.sdds.plasma.sd.service.styles
+package com.sdds.plasma.sd.service.styles.textfield
 
+import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.unit.dp
 import com.sdds.compose.uikit.ChipGroup
 import com.sdds.compose.uikit.EmbeddedChip
-import com.sdds.compose.uikit.TextAreaClear
+import com.sdds.compose.uikit.ScrollBar
+import com.sdds.compose.uikit.TextArea
 import com.sdds.compose.uikit.TextField
 import com.sdds.compose.uikit.TextFieldStyle
 import com.sdds.compose.uikit.TextFieldStyleBuilder
+import com.sdds.compose.uikit.adjustBy
+import com.sdds.plasma.sd.service.styles.chip.L
+import com.sdds.plasma.sd.service.styles.chip.M
+import com.sdds.plasma.sd.service.styles.chip.S
+import com.sdds.plasma.sd.service.styles.chip.Secondary
+import com.sdds.plasma.sd.service.styles.chip.Xs
+import com.sdds.plasma.sd.service.styles.chip.group.Dense
 import com.sdds.plasma.sd.service.tokens.compose.PlasmaSdServiceTheme
 
-val TextAreaClear.Xs: TextFieldStyleBuilder
+val TextArea.Xs: TextFieldStyleBuilder
     @Composable
     get() = TextFieldStyle.builder(this)
-        .fieldAppearance(TextField.FieldAppearance.Clear)
-        .helperTextPlacement(TextField.HelperTextPlacement.Outer)
+        .fieldAppearance(TextField.FieldAppearance.Solid)
+        .shape(PlasmaSdServiceTheme.shapes.roundS)
+        .helperTextPlacement(TextField.HelperTextPlacement.Inner)
         .chipGroupStyle(
             ChipGroup.Dense.chipStyle(
                 EmbeddedChip.Xs.Secondary.style(),
             ).style(),
         )
+        .scrollBar(scrollBar)
         .dimensions(
             TextField.Dimensions(
-                boxPaddingStart = 0.dp,
-                boxPaddingEnd = 0.dp,
+                boxPaddingStart = 8.dp,
+                boxPaddingEnd = 8.dp,
                 boxPaddingTopInnerLabel = 8.dp,
                 boxPaddingBottomInnerLabel = 8.dp,
                 boxPaddingTopOuterLabel = 8.dp,
                 boxPaddingBottomOuterLabel = 8.dp,
                 innerLabelPadding = 0.dp,
-                outerLabelPadding = 2.dp,
+                outerLabelPadding = 6.dp,
                 optionalPadding = 4.dp,
-                helperTextPaddingInner = 0.dp,
+                helperTextPaddingInner = 8.dp,
                 helperTextPaddingOuter = 4.dp,
                 startContentEndPadding = 4.dp,
                 endContentStartPadding = 4.dp,
@@ -44,10 +55,6 @@ val TextAreaClear.Xs: TextFieldStyleBuilder
                     startLabelVerticalPadding = 0.dp,
                     endLabelHorizontalPadding = 4.dp,
                     endLabelVerticalPadding = 2.dp,
-                    startFieldHorizontalPadding = 4.dp,
-                    startFieldVerticalPadding = 0.dp,
-                    endFieldHorizontalPadding = 4.dp,
-                    endFieldVerticalPadding = 0.dp,
                     fieldIndicatorSize = 6.dp,
                     labelIndicatorSize = 6.dp,
                 ),
@@ -64,28 +71,30 @@ val TextAreaClear.Xs: TextFieldStyleBuilder
         .dropInnerLabel(true)
         .singleLine(false)
 
-val TextAreaClear.S: TextFieldStyleBuilder
+val TextArea.S: TextFieldStyleBuilder
     @Composable
     get() = TextFieldStyle.builder(this)
-        .fieldAppearance(TextField.FieldAppearance.Clear)
-        .helperTextPlacement(TextField.HelperTextPlacement.Outer)
+        .fieldAppearance(TextField.FieldAppearance.Solid)
+        .shape(PlasmaSdServiceTheme.shapes.roundM.adjustBy(all = (-2).dp))
+        .helperTextPlacement(TextField.HelperTextPlacement.Inner)
         .chipGroupStyle(
             ChipGroup.Dense.chipStyle(
                 EmbeddedChip.S.Secondary.style(),
             ).style(),
         )
+        .scrollBar(scrollBar)
         .dimensions(
             TextField.Dimensions(
-                boxPaddingStart = 0.dp,
-                boxPaddingEnd = 0.dp,
+                boxPaddingStart = 12.dp,
+                boxPaddingEnd = 12.dp,
                 boxPaddingTopInnerLabel = 4.dp,
-                boxPaddingBottomInnerLabel = 4.dp,
+                boxPaddingBottomInnerLabel = 12.dp,
                 boxPaddingTopOuterLabel = 8.dp,
-                boxPaddingBottomOuterLabel = 8.dp,
+                boxPaddingBottomOuterLabel = 12.dp,
                 innerLabelPadding = 0.dp,
-                outerLabelPadding = 4.dp,
+                outerLabelPadding = 8.dp,
                 optionalPadding = 4.dp,
-                helperTextPaddingInner = 0.dp,
+                helperTextPaddingInner = 12.dp,
                 helperTextPaddingOuter = 4.dp,
                 startContentEndPadding = 4.dp,
                 endContentStartPadding = 6.dp,
@@ -94,14 +103,10 @@ val TextAreaClear.S: TextFieldStyleBuilder
                 alignmentLineHeight = 40.dp,
                 iconSize = 24.dp,
                 indicatorDimensions = TextField.Dimensions.IndicatorDimensions(
-                    startLabelHorizontalPadding = 6.dp,
+                    startLabelHorizontalPadding = 4.dp,
                     startLabelVerticalPadding = 0.dp,
                     endLabelHorizontalPadding = 4.dp,
                     endLabelVerticalPadding = 4.dp,
-                    startFieldHorizontalPadding = 6.dp,
-                    startFieldVerticalPadding = 0.dp,
-                    endFieldHorizontalPadding = 6.dp,
-                    endFieldVerticalPadding = 0.dp,
                     fieldIndicatorSize = 6.dp,
                     labelIndicatorSize = 6.dp,
                 ),
@@ -117,28 +122,30 @@ val TextAreaClear.S: TextFieldStyleBuilder
         .placeholderStyle(PlasmaSdServiceTheme.typography.bodySNormal)
         .singleLine(false)
 
-val TextAreaClear.M: TextFieldStyleBuilder
+val TextArea.M: TextFieldStyleBuilder
     @Composable
     get() = TextFieldStyle.builder(this)
-        .fieldAppearance(TextField.FieldAppearance.Clear)
-        .helperTextPlacement(TextField.HelperTextPlacement.Outer)
+        .fieldAppearance(TextField.FieldAppearance.Solid)
+        .shape(PlasmaSdServiceTheme.shapes.roundM)
+        .helperTextPlacement(TextField.HelperTextPlacement.Inner)
         .chipGroupStyle(
             ChipGroup.Dense.chipStyle(
                 EmbeddedChip.M.Secondary.style(),
             ).style(),
         )
+        .scrollBar(scrollBar)
         .dimensions(
             TextField.Dimensions(
-                boxPaddingStart = 0.dp,
-                boxPaddingEnd = 0.dp,
+                boxPaddingStart = 14.dp,
+                boxPaddingEnd = 14.dp,
                 boxPaddingTopInnerLabel = 6.dp,
-                boxPaddingBottomInnerLabel = 6.dp,
+                boxPaddingBottomInnerLabel = 12.dp,
                 boxPaddingTopOuterLabel = 12.dp,
                 boxPaddingBottomOuterLabel = 12.dp,
                 innerLabelPadding = 2.dp,
-                outerLabelPadding = 4.dp,
+                outerLabelPadding = 10.dp,
                 optionalPadding = 4.dp,
-                helperTextPaddingInner = 0.dp,
+                helperTextPaddingInner = 12.dp,
                 helperTextPaddingOuter = 4.dp,
                 startContentEndPadding = 6.dp,
                 endContentStartPadding = 8.dp,
@@ -151,12 +158,8 @@ val TextAreaClear.M: TextFieldStyleBuilder
                     startLabelVerticalPadding = 0.dp,
                     endLabelHorizontalPadding = 4.dp,
                     endLabelVerticalPadding = 4.dp,
-                    startFieldHorizontalPadding = 6.dp,
-                    startFieldVerticalPadding = 0.dp,
-                    endFieldHorizontalPadding = 6.dp,
-                    endFieldVerticalPadding = 0.dp,
                     fieldIndicatorSize = 8.dp,
-                    labelIndicatorSize = 8.dp,
+                    labelIndicatorSize = 6.dp,
                 ),
             ),
         )
@@ -170,28 +173,30 @@ val TextAreaClear.M: TextFieldStyleBuilder
         .placeholderStyle(PlasmaSdServiceTheme.typography.bodyMNormal)
         .singleLine(false)
 
-val TextAreaClear.L: TextFieldStyleBuilder
+val TextArea.L: TextFieldStyleBuilder
     @Composable
     get() = TextFieldStyle.builder(this)
-        .fieldAppearance(TextField.FieldAppearance.Clear)
-        .helperTextPlacement(TextField.HelperTextPlacement.Outer)
+        .fieldAppearance(TextField.FieldAppearance.Solid)
+        .shape(PlasmaSdServiceTheme.shapes.roundM.adjustBy(all = 2.dp))
+        .helperTextPlacement(TextField.HelperTextPlacement.Inner)
         .chipGroupStyle(
             ChipGroup.Dense.chipStyle(
                 EmbeddedChip.L.Secondary.style(),
             ).style(),
         )
+        .scrollBar(scrollBar)
         .dimensions(
             TextField.Dimensions(
-                boxPaddingStart = 0.dp,
-                boxPaddingEnd = 0.dp,
+                boxPaddingStart = 16.dp,
+                boxPaddingEnd = 16.dp,
                 boxPaddingTopInnerLabel = 9.dp,
-                boxPaddingBottomInnerLabel = 9.dp,
+                boxPaddingBottomInnerLabel = 12.dp,
                 boxPaddingTopOuterLabel = 16.dp,
-                boxPaddingBottomOuterLabel = 16.dp,
+                boxPaddingBottomOuterLabel = 12.dp,
                 innerLabelPadding = 2.dp,
-                outerLabelPadding = 4.dp,
+                outerLabelPadding = 12.dp,
                 optionalPadding = 4.dp,
-                helperTextPaddingInner = 0.dp,
+                helperTextPaddingInner = 12.dp,
                 helperTextPaddingOuter = 4.dp,
                 startContentEndPadding = 8.dp,
                 endContentStartPadding = 10.dp,
@@ -204,12 +209,8 @@ val TextAreaClear.L: TextFieldStyleBuilder
                     startLabelVerticalPadding = 0.dp,
                     endLabelHorizontalPadding = 4.dp,
                     endLabelVerticalPadding = 4.dp,
-                    startFieldHorizontalPadding = 6.dp,
-                    startFieldVerticalPadding = 0.dp,
-                    endFieldHorizontalPadding = 6.dp,
-                    endFieldVerticalPadding = 0.dp,
                     fieldIndicatorSize = 8.dp,
-                    labelIndicatorSize = 8.dp,
+                    labelIndicatorSize = 6.dp,
                 ),
             ),
         )
@@ -222,3 +223,12 @@ val TextAreaClear.L: TextFieldStyleBuilder
         .counterStyle(PlasmaSdServiceTheme.typography.bodyXsNormal)
         .placeholderStyle(PlasmaSdServiceTheme.typography.bodyLNormal)
         .singleLine(false)
+
+private val scrollBar: ScrollBar
+    @Composable
+    get() = ScrollBar(
+        indicatorThickness = 1.dp,
+        indicatorColor = PlasmaSdServiceTheme.colors.surfaceDefaultTransparentTertiary,
+        backgroundColor = PlasmaSdServiceTheme.colors.surfaceDefaultTransparentPrimary,
+        padding = PaddingValues(top = 18.dp, end = 2.dp, bottom = 36.dp),
+    )
