+++ tokens/plasma.sd.service.sandbox.compose/src/main/kotlin/com/sdds/plasma/sd/service/sandbox/SandboxActivity.kt
@@ -48,9 +48,9 @@ import com.sdds.plasma.sd.service.sandbox.radiobox.RadioBoxScreen
 import com.sdds.plasma.sd.service.sandbox.radiobox.group.RadioBoxGroupScreen
 import com.sdds.plasma.sd.service.sandbox.switch.SwitchScreen
 import com.sdds.plasma.sd.service.sandbox.textfield.TextFieldScreen
-import com.sdds.plasma.sd.service.styles.Clear
-import com.sdds.plasma.sd.service.styles.M
-import com.sdds.plasma.sd.service.styles.Pilled
+import com.sdds.plasma.sd.service.styles.button.icon.Clear
+import com.sdds.plasma.sd.service.styles.button.icon.M
+import com.sdds.plasma.sd.service.styles.button.icon.Pilled
 import com.sdds.plasma.sd.service.tokens.compose.PlasmaSdServiceTheme
 import kotlinx.coroutines.launch
 import com.sdds.icons.R.drawable as Icons
