+++ tokens/plasma.sd.service.sandbox.compose/src/main/kotlin/com/sdds/plasma/sd/service/sandbox/Theme.kt
@@ -37,11 +37,24 @@ import com.sdds.compose.uikit.Switch
 import com.sdds.compose.uikit.TextField
 import com.sdds.compose.uikit.internal.focusselector.FocusSelectorMode
 import com.sdds.compose.uikit.internal.focusselector.LocalFocusSelectorMode
-import com.sdds.plasma.sd.service.styles.Default
 import com.sdds.plasma.sd.service.styles.M
-import com.sdds.plasma.sd.service.styles.Optional
-import com.sdds.plasma.sd.service.styles.OuterLabel
-import com.sdds.plasma.sd.service.styles.S
+import com.sdds.plasma.sd.service.styles.avatar.M
+import com.sdds.plasma.sd.service.styles.avatar.S
+import com.sdds.plasma.sd.service.styles.button.basic.Default
+import com.sdds.plasma.sd.service.styles.button.basic.M
+import com.sdds.plasma.sd.service.styles.button.icon.Default
+import com.sdds.plasma.sd.service.styles.button.icon.M
+import com.sdds.plasma.sd.service.styles.checkbox.M
+import com.sdds.plasma.sd.service.styles.checkbox.group.M
+import com.sdds.plasma.sd.service.styles.chip.Default
+import com.sdds.plasma.sd.service.styles.chip.M
+import com.sdds.plasma.sd.service.styles.progress.Default
+import com.sdds.plasma.sd.service.styles.radiobox.M
+import com.sdds.plasma.sd.service.styles.radiobox.group.M
+import com.sdds.plasma.sd.service.styles.textfield.Default
+import com.sdds.plasma.sd.service.styles.textfield.M
+import com.sdds.plasma.sd.service.styles.textfield.Optional
+import com.sdds.plasma.sd.service.styles.textfield.OuterLabel
 import com.sdds.plasma.sd.service.tokens.compose.PlasmaSdServiceTheme
 import com.sdds.plasma.sd.service.tokens.compose.darkPlasmaSdServiceColors
 import com.sdds.plasma.sd.service.tokens.compose.darkPlasmaSdServiceGradients
