+++ tokens/plasma.sd.service.sandbox.compose/src/main/kotlin/com/sdds/plasma/sd/service/sandbox/avatar/AvatarGroupScreen.kt
@@ -16,7 +16,7 @@ import com.sdds.compose.uikit.Image
 import com.sdds.compose.uikit.avatar
 import com.sdds.plasma.sd.service.sandbox.compose.R
 import com.sdds.plasma.sd.service.sandbox.core.ComponentScaffold
-import com.sdds.plasma.sd.service.styles.S
+import com.sdds.plasma.sd.service.styles.avatar.S
 import com.sdds.plasma.sd.service.tokens.compose.PlasmaSdServiceTheme
 import com.sdds.icons.R.drawable as Icons
