+++ tokens/plasma.sd.service.sandbox.compose/src/main/kotlin/com/sdds/plasma/sd/service/sandbox/buttons/ButtonUiState.kt
@@ -9,21 +9,49 @@ import com.sdds.compose.uikit.IconButton
 import com.sdds.compose.uikit.IconButtonStyleBuilder
 import com.sdds.compose.uikit.LinkButton
 import com.sdds.compose.uikit.LinkButtonStyleBuilder
-import com.sdds.plasma.sd.service.styles.Accent
-import com.sdds.plasma.sd.service.styles.Black
-import com.sdds.plasma.sd.service.styles.Clear
-import com.sdds.plasma.sd.service.styles.Dark
-import com.sdds.plasma.sd.service.styles.Default
-import com.sdds.plasma.sd.service.styles.L
-import com.sdds.plasma.sd.service.styles.M
-import com.sdds.plasma.sd.service.styles.Negative
-import com.sdds.plasma.sd.service.styles.Pilled
-import com.sdds.plasma.sd.service.styles.Positive
-import com.sdds.plasma.sd.service.styles.S
-import com.sdds.plasma.sd.service.styles.Secondary
-import com.sdds.plasma.sd.service.styles.Warning
-import com.sdds.plasma.sd.service.styles.White
-import com.sdds.plasma.sd.service.styles.Xs
+import com.sdds.plasma.sd.service.styles.button.basic.Accent
+import com.sdds.plasma.sd.service.styles.button.basic.Black
+import com.sdds.plasma.sd.service.styles.button.basic.Clear
+import com.sdds.plasma.sd.service.styles.button.basic.Dark
+import com.sdds.plasma.sd.service.styles.button.basic.Default
+import com.sdds.plasma.sd.service.styles.button.basic.L
+import com.sdds.plasma.sd.service.styles.button.basic.M
+import com.sdds.plasma.sd.service.styles.button.basic.Negative
+import com.sdds.plasma.sd.service.styles.button.basic.Positive
+import com.sdds.plasma.sd.service.styles.button.basic.S
+import com.sdds.plasma.sd.service.styles.button.basic.Secondary
+import com.sdds.plasma.sd.service.styles.button.basic.Warning
+import com.sdds.plasma.sd.service.styles.button.basic.White
+import com.sdds.plasma.sd.service.styles.button.basic.Xs
+import com.sdds.plasma.sd.service.styles.button.icon.Accent
+import com.sdds.plasma.sd.service.styles.button.icon.Black
+import com.sdds.plasma.sd.service.styles.button.icon.Clear
+import com.sdds.plasma.sd.service.styles.button.icon.Dark
+import com.sdds.plasma.sd.service.styles.button.icon.Default
+import com.sdds.plasma.sd.service.styles.button.icon.L
+import com.sdds.plasma.sd.service.styles.button.icon.M
+import com.sdds.plasma.sd.service.styles.button.icon.Negative
+import com.sdds.plasma.sd.service.styles.button.icon.Pilled
+import com.sdds.plasma.sd.service.styles.button.icon.Positive
+import com.sdds.plasma.sd.service.styles.button.icon.S
+import com.sdds.plasma.sd.service.styles.button.icon.Secondary
+import com.sdds.plasma.sd.service.styles.button.icon.Warning
+import com.sdds.plasma.sd.service.styles.button.icon.White
+import com.sdds.plasma.sd.service.styles.button.icon.Xs
+import com.sdds.plasma.sd.service.styles.button.link.Accent
+import com.sdds.plasma.sd.service.styles.button.link.Black
+import com.sdds.plasma.sd.service.styles.button.link.Clear
+import com.sdds.plasma.sd.service.styles.button.link.Dark
+import com.sdds.plasma.sd.service.styles.button.link.Default
+import com.sdds.plasma.sd.service.styles.button.link.L
+import com.sdds.plasma.sd.service.styles.button.link.M
+import com.sdds.plasma.sd.service.styles.button.link.Negative
+import com.sdds.plasma.sd.service.styles.button.link.Positive
+import com.sdds.plasma.sd.service.styles.button.link.S
+import com.sdds.plasma.sd.service.styles.button.link.Secondary
+import com.sdds.plasma.sd.service.styles.button.link.Warning
+import com.sdds.plasma.sd.service.styles.button.link.White
+import com.sdds.plasma.sd.service.styles.button.link.Xs
 import com.sdds.icons.R.drawable as Icons
 
 /**
