+++ tokens/plasma.sd.service.sandbox.compose/src/main/kotlin/com/sdds/plasma/sd/service/sandbox/checkbox/group/SandboxCheckBoxGroup.kt
@@ -6,7 +6,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import com.sdds.compose.uikit.CheckBox
 import com.sdds.compose.uikit.CheckBoxGroup
 import com.sdds.plasma.sd.service.sandbox.SandboxTheme
-import com.sdds.plasma.sd.service.styles.M
+import com.sdds.plasma.sd.service.styles.checkbox.group.M
 
 /**
  * Превью [CheckBoxGroup]
