+++ tokens/plasma.sd.service.sandbox.compose/src/main/kotlin/com/sdds/plasma/sd/service/sandbox/core/Editors.kt
@@ -28,8 +28,8 @@ import com.sdds.compose.uikit.RadioBox
 import com.sdds.compose.uikit.Text
 import com.sdds.compose.uikit.TextField
 import com.sdds.plasma.sd.service.sandbox.compose.R
-import com.sdds.plasma.sd.service.styles.Default
-import com.sdds.plasma.sd.service.styles.M
+import com.sdds.plasma.sd.service.styles.button.basic.Default
+import com.sdds.plasma.sd.service.styles.button.basic.M
 import com.sdds.plasma.sd.service.tokens.compose.PlasmaSdServiceTheme
 import kotlin.math.roundToInt
