+++ tokens/plasma.sd.service.sandbox.compose/src/main/kotlin/com/sdds/plasma/sd/service/sandbox/progress/ProgressUiState.kt
@@ -3,13 +3,13 @@ package com.sdds.plasma.sd.service.sandbox.progress
 import androidx.compose.runtime.Composable
 import com.sdds.compose.uikit.ProgressBar
 import com.sdds.compose.uikit.ProgressBarStyle
-import com.sdds.plasma.sd.service.styles.Accent
-import com.sdds.plasma.sd.service.styles.Default
-import com.sdds.plasma.sd.service.styles.GradientAccent
-import com.sdds.plasma.sd.service.styles.Negative
-import com.sdds.plasma.sd.service.styles.Positive
-import com.sdds.plasma.sd.service.styles.Secondary
-import com.sdds.plasma.sd.service.styles.Warning
+import com.sdds.plasma.sd.service.styles.progress.Accent
+import com.sdds.plasma.sd.service.styles.progress.Default
+import com.sdds.plasma.sd.service.styles.progress.GradientAccent
+import com.sdds.plasma.sd.service.styles.progress.Negative
+import com.sdds.plasma.sd.service.styles.progress.Positive
+import com.sdds.plasma.sd.service.styles.progress.Secondary
+import com.sdds.plasma.sd.service.styles.progress.Warning
 
 /**
  * Состояние компонента Progress
